update docusaurus from 2.0.0-alpha.70 to 2.0.0-beta.1 (#4889)
* exit with an error if no paths given to copy script

* consolidate and normalize scripts in website package

* update docusaurus from 2.0.0-alpha.70 to 2.0.0-beta.1

* fix badge url in fast-components doc

+++ build/clean.js
@@ -1,3 +1,5 @@
+/* eslint-env node */
+/* eslint-disable @typescript-eslint/explicit-function-return-type,@typescript-eslint/no-var-requires,@typescript-eslint/typedef */
 /**
  * Utility for cleaning directories.
  * Usage: node build/clean.js %path%
@@ -15,7 +17,11 @@ const paths = argv._;
  * Function to remove a given path
  */
 function cleanPath(cleanPath) {
-    const removePath = path.resolve(process.cwd(), cleanPath)
+    if (!cleanPath) {
+        console.error("No path specified.");
+        process.exit(1);
+    }
+    const removePath = path.resolve(process.cwd(), cleanPath);
     rimraf(removePath, () => {
         console.log(removePath, "cleaned");
     });
